Remove fakeTarget
Instead of creating a "fake" target, just create a shallow clone of the
source object immediately and use that as the target, since we have to
clone it anyway once any modifications are made. The only case in which
this might be slower is if the callback makes no modifications at all,
but that seems unlikely given the purpose of this library.